VR demo behavior option added, not yet implemented; sum added to countly events

+++ src/edu/usf/eng/pie/avatars4change/userData/userData.java
@@ -6,7 +6,7 @@ public class userData {
     
     //user context:
     public static String currentActivity = "???";
-    public static int    currentActivityLevel = 0;
+    public static float    currentActivityLevel = 0;
     public static double[] FFT = new double[65];
 	
 }
